+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   this made it impossible to load files/thumbnails via `<img>` tag or the like
   and the long hash alone should be enough to make those routes nearly
   impossible to guess
++ Made current password a requirement when updating user
 + Updated dependencies
 
 ## [1.4.0] - 2018-05-13